api/face-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.6
insightface==0.7.3
onnxruntime>=1.16.0
opencv-python-headless>=4.9.0.80
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
